py_backend/requirements.txt
# Web framework
fastapi
uvicorn
python-multipart  # For file uploads
pydantic # Using 1.x for better compatibility

# LangChain and related
langchain
langchain-community
langchain-core
chromadb
langchain-chroma
langchain-ollama

# Document processing
pypdf
python-docx
pandas

# Utilities
requests

# IMPORTANT: Keep numpy locked at a compatible version
numpy

faster-whisper

diffusers[torch]
torch
transformers
safetensors
accelerate

# Keycloak integration
python-keycloak
python-jose[cryptography]
